[CCFPCM-395] mock data moved to one file - fixed imports throughout
