+++ apps/backend/src/deposits/entities/cash-deposit.entity.ts
@@ -5,10 +5,10 @@ import {
   Entity,
   PrimaryGeneratedColumn
 } from 'typeorm';
-import { FileTypes } from './../../constants';
 import { FileMetadata } from '../../common/columns/metadata';
 import { MatchStatus } from '../../common/const';
 import { ColumnNumericTransformer } from '../../common/transformers/numericColumnTransformer';
+import { FileTypes } from '../../constants';
 import { TDI17Details } from '../../flat-files';
 import { PaymentEntity } from '../../transaction/entities/payment.entity';
