+++ apps/backend/src/reconciliation/cash-exceptions.service.ts
@@ -2,11 +2,11 @@ import { Injectable, Inject, Logger } from '@nestjs/common';
 import { MatchStatus } from '../common/const';
 import { Ministries } from '../constants';
 import { CashDepositService } from '../deposits/cash-deposit.service';
+import { CashDepositEntity } from '../deposits/entities/cash-deposit.entity';
 import { LocationEntity } from '../location/entities';
 import { AppLogger } from '../logger/logger.service';
 import { PaymentEntity } from '../transaction/entities/payment.entity';
 import { PaymentService } from '../transaction/payment.service';
-import { CashDepositEntity } from './../deposits/entities/cash-deposit.entity';
 
 @Injectable()
 export class CashExceptionsService {
